+++ includes/features/class-events.php
@@ -662,7 +662,7 @@ class Events extends \WP_List_Table {
 					} else {
 						$local = false;
 					}
-					if ( Role::SUPER_ADMIN === Role::admin_type() || Role::SINGLE_ADMIN === Role::admin_type() || ( Role::LOCAL_ADMIN === Role::admin_type() && $local ) ) {
+					if ( Role::SUPER_ADMIN === Role::admin_type() || Role::SINGLE_ADMIN === Role::admin_type() || ( Role::LOCAL_ADMIN === Role::admin_type() && $local ) || Role::override_privileges() ) {
 						$log = [
 							'name'    => $archiver['name'],
 							'running' => $archiver['running'],